Create common ReferenceDataSwitch component
